refactor: destroy category
