+++ src/actions/cms/articleCategories/destroy.ts
@@ -3,21 +3,9 @@ import prisma from '@/prisma'
 import { createPrismaActionError } from '@/actions/error'
 import type { ExpandedArticleCategory } from '@/cms/articleCategories/Types'
 import type { ActionReturn } from '@/actions/Types'
+import { destroyArticleCategory } from '@/server/cms/articleCategories/destroy'
 
-export async function destroyArticleCategory(id: number): Promise<ActionReturn<ExpandedArticleCategory>> {
-    try {
-        // TODO: Cheek for visibility type edit of user.
-
-        const articleCategory = await prisma.articleCategory.delete({
-            where: {
-                id
-            },
-            include: {
-                articles: true
-            }
-        })
-        return { success: true, data: articleCategory }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+export async function destroyArticleCategoryAction(id: number): Promise<ActionReturn<ExpandedArticleCategory>> {
+    // TODO: Cheek for visibility type edit of user.
+    return await destroyArticleCategory(id)
 }
